TASK: Move EventListenerManager::catchup to Publisher
